refactor: extracted quick edit to separate component logic
